+++ UserFunction.hpp
@@ -21,7 +21,7 @@ public:
   virtual NodePtr EvalNum(const std::vector<NumberRepr> &args) = 0;
   virtual bool SpecialValues(const std::list<NodePtr> &args,
                              NodePtr *result) = 0;
-  virtual int NumArgs() const = 0;
+  virtual size_t NumArgs() const = 0;
 
   static NodePtr make_node(const std::string &expr);
 };
@@ -30,7 +30,7 @@ class FuncSin : public UserFunction {
 public:
   virtual NodePtr EvalNum(const std::vector<NumberRepr> &args);
   virtual bool SpecialValues(const std::list<NodePtr> &args, NodePtr *result);
-  virtual int NumArgs() const { return 1; }
+  virtual size_t NumArgs() const { return 1; }
 
   static std::list<std::pair<NodePtr, std::string>> svalues;
 };
@@ -39,7 +39,7 @@ class FuncCos : public UserFunction {
 public:
   virtual NodePtr EvalNum(const std::vector<NumberRepr> &args);
   virtual bool SpecialValues(const std::list<NodePtr> &args, NodePtr *result);
-  virtual int NumArgs() const { return 1; }
+  virtual size_t NumArgs() const { return 1; }
 
   static std::list<std::pair<NodePtr, std::string>> svalues;
 };
